requirements.txt
drf-yasg==1.21.5
django-cors-headers==4.1.0
Django==4.2.20
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
django-redis==5.4.0
channels-redis==4.1.0
channels==4.0.0
celery==5.3.6
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.4.1 